fix int overflows

+++ android/src/main/java/com/reactnativestripesdk/FinancialConnectionsSheetFragment.kt
@@ -123,7 +123,7 @@ class FinancialConnectionsSheetFragment : Fragment() {
       map.putString("status", account.status.value)
       map.putString("institutionName", account.institutionName)
       map.putString("last4", account.last4)
-      map.putInt("created", account.created * 1000)
+      map.putDouble("created", account.created * 1000.0)
       map.putMap("balance", mapFromAccountBalance(account.balance))
       map.putMap("balanceRefresh", mapFromAccountBalanceRefresh(account.balanceRefresh))
       map.putString("category", account.category.value)
@@ -140,7 +140,7 @@ class FinancialConnectionsSheetFragment : Fragment() {
       return null
     }
     val map = WritableNativeMap()
-    map.putInt("asOf", balance.asOf * 1000)
+    map.putDouble("asOf", balance.asOf * 1000.0)
     map.putString("type", balance.type.value)
     map.putMap("current", balance.current as ReadableMap)
     WritableNativeMap().also {
@@ -160,7 +160,7 @@ class FinancialConnectionsSheetFragment : Fragment() {
     }
     val map = WritableNativeMap()
     map.putString("status", balanceRefresh.status?.name) // TODO check if this is the correct or if we need raw value
-    map.putInt("lastAttemptedAt", balanceRefresh.lastAttemptedAt * 1000)
+    map.putDouble("lastAttemptedAt", balanceRefresh.lastAttemptedAt * 1000.0)
     return map
   }
 }
